# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exam Result System API",
        "description": "API documentation for the high-performance Exam Result System.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/examinations": {
            "get": {
                "tags": [
                    "Examinations"
                ],
                "summary": "List all examinations",
                "operationId": "8599f4a9bfb42dbfb86093c76d690cc8",
                "responses": {
                    "200": {
                        "description": "Paginated list of examinations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Examinations"
                ],
                "summary": "Create a new examination",
                "operationId": "49fcde738a38309641c0d27032183ab7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "programme_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Fall 2026"
                                    },
                                    "academic_year": {
                                        "type": "string",
                                        "example": "2026-2027"
                                    },
                                    "term": {
                                        "type": "string",
                                        "example": "Fall"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Examination created"
                    }
                }
            }
        },
        "/api/examinations/{examination}": {
            "get": {
                "tags": [
                    "Examinations"
                ],
                "summary": "Show an examination with its courses and components",
                "operationId": "a4b41b2505097eddd0c2ab700792df5c",
                "parameters": [
                    {
                        "name": "examination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Examination details"
                    }
                }
            }
        },
        "/api/examinations/{examination}/status": {
            "patch": {
                "tags": [
                    "Examinations"
                ],
                "summary": "Update the examination lifecycle status",
                "operationId": "669d66608b8ff2ba643d43c010574d11",
                "parameters": [
                    {
                        "name": "examination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "open",
                                            "locked",
                                            "computing",
                                            "published"
                                        ],
                                        "example": "open"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                }
            }
        },
        "/api/examinations/{examination}/courses": {
            "post": {
                "tags": [
                    "Examinations"
                ],
                "summary": "Attach a course to an examination",
                "operationId": "25a2962fad65e24dc4feb0de4535379f",
                "parameters": [
                    {
                        "name": "examination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "max_marks": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "pass_marks": {
                                        "type": "integer",
                                        "example": 40
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course attached to examination"
                    }
                }
            }
        },
...
        "/api/examinations/{examination}/marks/upload": {
            "post": {
                "tags": [
                    "Marks"
                ],
                "summary": "Bulk upload marks via CSV",
                "operationId": "45487cfd94f444af144d134390a11682",
                "parameters": [
                    {
                        "name": "examination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Batch job queued"
                    }
                }
            }
        },
        "/api/mark-uploads/{batch}": {
            "get": {
                "tags": [
                    "Marks"
                ],
                "summary": "Check status of a bulk upload batch",
                "operationId": "9fb6a74d8508460d59c048f71d91f3d4",
                "parameters": [
                    {
                        "name": "batch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch status"
                    }
                }
            }
        },
        "/api/mark-uploads/{batch}/errors": {
            "get": {
                "tags": [
                    "Marks"
                ],
                "summary": "Get paginated row-level import errors for a batch",
                "operationId": "f04d3566f062f0cd8dd1cbd0122f2f4e",
                "parameters": [
                    {
                        "name": "batch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of row-level errors"
                    }
                }
            }
        },
        "/api/programmes": {
            "get": {
                "tags": [
                    "Programmes"
                ],
                "summary": "List all programmes",
                "operationId": "b89777fa35530f2a15cb6f0d7c91dfbd",
                "responses": {
                    "200": {
                        "description": "Paginated list of programmes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Programmes"
                ],
                "summary": "Create a new programme",
                "operationId": "b9efff61fead5548f41d4a916b0db453",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "CSE"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Computer Science & Engineering"
                                    },
                                    "duration_years": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Programme created"
                    }
                }
            }
        },
        "/api/programmes/{programme}": {
            "get": {
                "tags": [
                    "Programmes"
                ],
                "summary": "Get a specific programme with its courses",
                "operationId": "9e55031a2bfd1fc157443b33e7d5929d",
                "parameters": [
                    {
                        "name": "programme",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Programme details"
                    }
                }
            }
        },
...
    },
...
}